+++ packages/blocks/core/js/wordpress/query-title/test/functionality.blocks-2.e2e.cy.js
@@ -42,7 +42,20 @@ describe('Query Title Block', () => {
 		);
 
 		//
-		// 2. Assert inner blocks selectors in front end
+		// 2. Check settings tab
+		//
+		cy.getByDataTest('settings-tab').click();
+
+		// layout settings should be hidden
+		cy.get('.block-editor-block-inspector').within(() => {
+			cy.get('.components-tools-panel-header')
+				.contains('Settings')
+				.scrollIntoView()
+				.should('be.visible');
+		});
+
+		//
+		// 3. Assert inner blocks selectors in front end
 		//
 
 		// it's very complicated to test this, because the block is not rendered in the front end