+++ packages/blocks/core/CHANGELOG.md
@@ -113,6 +113,7 @@
 - `Post Title` block: support improved.
 - `Query Pagination` block: support improved.
 - `Query Pagination Numbers` block: support improved.
+- `Query Title` block: support improved.
 
 
 ### Automated Tests
@@ -163,6 +164,7 @@
 - Updated E2E tests to check support of `Post Title` block.
 - Updated E2E tests to check support of `Query Pagination` block.
 - Updated E2E tests to check support of `Query Pagination Numbers` block.
+- Updated E2E tests to check support of `Query Title` block.
 
 
 ## 1.3.0 (2025-04-27)